Renamed to ResourceContainerRepository
