+++ src/main/kotlin/org/wycliffeassociates/otter/common/persistence/repositories/IResourceContainerRepository.kt
@@ -1,9 +1,9 @@
-package org.wycliffeassociates.otter.common.persistence
+package org.wycliffeassociates.otter.common.persistence.repositories
 
 import io.reactivex.Completable
 import org.wycliffeassociates.otter.common.collections.tree.Tree
 import org.wycliffeassociates.resourcecontainer.ResourceContainer
 
-interface IResourceContainerTreeImporter {
+interface IResourceContainerRepository {
     fun importResourceContainer(rc: ResourceContainer, rcTree: Tree, languageSlug: String): Completable
 }
